Track object file dependencies for TH accurately (#20604)
`hscCompileCoreExprHook` is changed to return a list of `Module`s required
by a splice. These modules are accumulated in the TcGblEnv (tcg_th_needed_mods).
Object file dependencies are added to the transitive closure set of these modules.

`getLinkDeps` is replaced by a simpler function (loadModules) that loads modules
without precomputing a transitive closure set.

The data structures in `LoaderState` are replaced with more efficient versions
to keep track of all the information required.